requirements.txt
gunicorn==23.0.0
gevent==24.11.1
requests==2.32.3
python-dotenv==1.0.1
pillow==11.1.0
redis==5.2.1
pymupdf4llm==0.0.17
DBUtils==3.1.0
pdfplumber==0.11.5
boto3==1.37.1
aioboto3==14.1.0
aiofiles==24.1.0
openai==1.66.3
beautifulsoup4==4.13.3
google-cloud-aiplatform==1.81.0
google-genai==1.11.0
google-cloud-vision==3.10.0
scikit-learn==1.6.1
opencv-python==4.11.0.86
pypdf==5.3.0
PyMySQL==1.1.1
python-box==7.3.2

# DB
SQLAlchemy==2.0.38
asyncmy===0.2.10
# db 동기화
alembic==1.14.1
# fastApi
fastapi==0.115.8
python-multipart==0.0.20

# task 실행
invoke==2.2.0

# debug
watchdog==6.0.0

# 서버 워커
uvicorn==0.34.0

# sentry
sentry_sdk===2.22.0

pandas==2.2.3

# gps
piexif==1.1.3
googlemaps==4.10.0

replicate==1.0.7

# gradio
gradio==5.44.1